fix show tickets activity, add scroll to show ticket layout
