+++ app/src/main/res/layout/bought_ticket_row.xml
@@ -10,7 +10,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_gravity="center_horizontal"
-        android:layout_margin="@dimen/material_default_padding"
+        android:layout_margin="@dimen/material_small_padding"
         android:textStyle="bold"
         android:textColor="@color/black"
         android:textAppearance="@style/TextAppearance.AppCompat.Medium"
